+++ src/internal.rs
@@ -25,32 +25,32 @@ pub fn readable_size(bytes: usize) -> (f64, &'static str) {
 }
 
 #[cfg(target_os = "linux")]
-pub fn peak_memory_usage() -> Result<usize, &'static str> {
+pub fn peak_memory_usage() -> Result<usize, String> {
     unsafe {
         let mut rusage: libc::rusage = std::mem::zeroed();
         let retval = libc::getrusage(libc::RUSAGE_SELF, &mut rusage as *mut _);
         match retval {
             0 => Ok(rusage.ru_maxrss as usize * 1024),
-            _ => Err("libc::getrusage call failed"),
+            val => Err(format!("libc::getrusage call failed with return value {}", val)),
         }
     }
 }
 
 #[cfg(target_os = "macos")]
-pub fn peak_memory_usage() -> Result<usize, &'static str> {
+pub fn peak_memory_usage() -> Result<usize, String> {
     unsafe {
         let mut rusage: libc::rusage = std::mem::zeroed();
         let retval = libc::getrusage(libc::RUSAGE_SELF, &mut rusage as *mut _);
         match retval {
             0 => Ok(rusage.ru_maxrss as usize),
-            _ => Err("libc::getrusage call failed"),
+            val => Err(format!("libc::getrusage call failed with return value {}", val)),
         }
     }
 }
 
 #[cfg(not(any(target_os = "linux", target_os = "macos")))]
-pub fn peak_memory_usage() -> Result<usize, &'static str> {
-    Err("No peak_memory_usage implementation for this OS")
+pub fn peak_memory_usage() -> Result<usize, String> {
+    Err(String::from("No peak_memory_usage implementation for this OS"))
 }
 
 //-----------------------------------------------------------------------------